Move the platform-specific stuff out of `Selfie.kt` and into `SnapshotStorage.kt`.
